requirements.txt
PySide6

# Data & analytics
pandas
numpy
matplotlib

# Génération & médias
python-barcode==0.15.1
Pillow==10.2.0
reportlab==4.1.0
qrcode==7.4.2

# Excel
openpyxl==3.1.2

# Cloud / Firebase
firebase-admin==6.5.0
google-cloud-firestore==2.11.3